Move FC *Impl types to config.flexible.internal package
