+++ engine/src/main/java/org/terasology/config/flexible/internal/SettingImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 MovingBlocks
+ * Copyright 2019 MovingBlocks
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.config.flexible;
+package org.terasology.config.flexible.internal;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Sets;
@@ -22,6 +22,7 @@ import com.google.gson.GsonBuilder;
 import com.google.gson.JsonElement;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.terasology.config.flexible.Setting;
 import org.terasology.config.flexible.constraints.SettingConstraint;
 import org.terasology.engine.SimpleUri;
 
@@ -35,7 +36,7 @@ import java.util.Set;
  *
  * @param <T> The type of the value this {@link SettingImpl} contains.
  */
-public class SettingImpl<T> implements Setting<T> {
+class SettingImpl<T> implements Setting<T> {
     private static final Logger LOGGER = LoggerFactory.getLogger(SettingImpl.class);
     private static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
