+++ engine/src/main/java/org/terasology/engine/subsystem/common/ConfigurationSubsystem.java
@@ -23,10 +23,8 @@ import org.terasology.config.facade.BindsConfiguration;
 import org.terasology.config.facade.BindsConfigurationImpl;
 import org.terasology.config.facade.InputDeviceConfiguration;
 import org.terasology.config.facade.InputDeviceConfigurationImpl;
-import org.terasology.config.flexible.FlexibleConfig;
-import org.terasology.config.flexible.FlexibleConfigImpl;
 import org.terasology.config.flexible.FlexibleConfigManager;
-import org.terasology.config.flexible.FlexibleConfigManagerImpl;
+import org.terasology.config.flexible.internal.FlexibleConfigManagerImpl;
 import org.terasology.context.Context;
 import org.terasology.engine.SimpleUri;
 import org.terasology.engine.TerasologyConstants;
